search-service/requirements.txt
# Search Service dependencies

# API framework
fastapi==0.115.11
uvicorn==0.29.0
pydantic==2.10.6
python-dotenv==1.0.1

# LangChain / LangGraph
langchain==0.3.21
langchain_community==0.3.20
langchain_core==0.3.46
langchain_neo4j==0.4.0
langchain_openai==0.3.9
langgraph==0.3.18

# Neo4j
neo4j>=5.0.0

# Data
numpy==1.26.2
pandas==2.2.3

# Embeddings
sentence_transformers==4.1.0
faiss-cpu==1.9.0
tiktoken
